backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
